Remove char* variants of QDataStream::operator<</>>
The correct way to use operator>> is to pass a char* reference to
get the data back, we cannot do that. also, we cannot prevent
ruby strings to be cast to const char * for operator<<, so we disable
char * and use the QString. This worked before because global
operators were prefered.